requirements.txt
# ==================== BossJy-Pro Enterprise Dependencies ====================
# Compatible: Python 3.11.9 (Recommended)
# Updated: 2025-10-08

# ==================== Core Web Frameworks ====================
fastapi==0.115.11
uvicorn[standard]==0.32.0
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.5.3
flask-socketio==5.3.5
Flask-Compress==1.14
python-socketio==5.10.0
eventlet==0.33.3
gunicorn==21.2.0

# ==================== Data Validation ====================
pydantic==2.10.6
pydantic-settings==2.11.0

# ==================== Database ====================
sqlalchemy==2.0.23
psycopg2-binary==2.9.10
psycopg==3.2.6
alembic==1.13.1

# ==================== Cache & Queue ====================
redis==5.3.1
fakeredis==2.31.3
celery==5.4.0
flower==2.0.1

# ==================== Telegram Bot ====================
python-telegram-bot==22.5

# ==================== HTTP Clients & Async ====================
httpx==0.27.0
requests==2.31.0
aiofiles==23.2.1
aiohttp==3.9.1
httptools==0.6.4
watchfiles==0.24.0

# ==================== File Processing ====================
pandas==2.2.2
numpy>=1.26.4,<2.0.0
openpyxl==3.1.2
xlrd==2.0.1
chardet==5.2.0

# ==================== AI & ML ====================
sentence-transformers==2.3.1
torch==2.2.2
transformers==4.38.0
google-generativeai==0.8.3
matplotlib==3.8.2

# ==================== Chinese Name Detection ====================
zhon==2.0.2
pypinyin==0.51.0
opencc-python-reimplemented==0.1.7
ethnicolr==0.18.4   # Ethnicity prediction

# ==================== International Name Detection ====================
nameparser==1.1.3

# ==================== Phone Number Validation ====================
phonenumbers==8.13.45
rapidfuzz==3.10.1

# ==================== Cryptocurrency Address Validation ====================
base58==2.1.1

# ==================== Security & Auth ====================
python-jose>=3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.5
pyjwt>=2.9.0,<3.0


# ==================== Utilities ====================
python-dotenv==1.0.0
python-multipart==0.0.9
pyyaml==6.0.1
python-dateutil==2.8.2

# ==================== Monitoring & Performance ====================
psutil==5.9.8
rich==13.7.0
prometheus-client==0.19.0

# ==================== Development & Testing ====================
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
flake8==6.1.0

# ==================== Logging ====================
loguru==0.7.2
colorlog==6.8.0

# ==================== Data Validation & Type Checking ====================
email-validator==2.1.0
validators==0.22.0
cerberus==1.3.7

# ==================== Address Parsing ====================
geopy==2.4.1
usaddress==0.5.16

# ==================== Google Cloud Services ====================
google-cloud-language==2.17.2
